+++ src/test/java/com/samhap/kokomen/interview/docs/InterviewDocsV2Test.java
@@ -29,7 +29,7 @@ import com.samhap.kokomen.interview.entity.RootQuestion;
 import com.samhap.kokomen.interview.repository.InterviewRepository;
 import com.samhap.kokomen.interview.repository.QuestionRepository;
 import com.samhap.kokomen.interview.repository.RootQuestionRepository;
-import com.samhap.kokomen.interview.service.InterviewFacadeService;
+import com.samhap.kokomen.interview.service.InterviewProceedFacadeService;
 import com.samhap.kokomen.interview.service.dto.AnswerRequestV2;
 import com.samhap.kokomen.member.domain.Member;
 import com.samhap.kokomen.member.repository.MemberRepository;
@@ -75,7 +75,7 @@ class InterviewDocsV2Test extends DocsTest {
         Question question3 = questionRepository.save(
                 QuestionFixtureBuilder.builder().interview(interview).content(rootQuestion.getContent()).build());
         answerRepository.save(AnswerFixtureBuilder.builder().question(question3).build());
-        String interviewProceedStateKey = InterviewFacadeService.createInterviewProceedStateKey(interview.getId(),
+        String interviewProceedStateKey = InterviewProceedFacadeService.createInterviewProceedStateKey(interview.getId(),
                 question3.getId());
         redisService.setValue(interviewProceedStateKey, InterviewProceedState.COMPLETED.name(), Duration.ofSeconds(10));
 
@@ -126,7 +126,7 @@ class InterviewDocsV2Test extends DocsTest {
         answerRepository.save(AnswerFixtureBuilder.builder().question(question2).build());
         questionRepository.save(
                 QuestionFixtureBuilder.builder().interview(interview).content(rootQuestion.getContent()).build());
-        String interviewProceedStateKey = InterviewFacadeService.createInterviewProceedStateKey(interview.getId(),
+        String interviewProceedStateKey = InterviewProceedFacadeService.createInterviewProceedStateKey(interview.getId(),
                 question2.getId());
         redisService.setValue(interviewProceedStateKey, InterviewProceedState.COMPLETED.name(), Duration.ofSeconds(10));
